// src/types/thumbnail.ts
export interface ThumbnailData {
    title: string;
    titleColor: string;
    titleFontSize: number;
    titleFontStyle: string;
    titleSpacing: number;
    bgImage: string;
    fgImage: string;
    fgPosition: {
      x: number;
      y: number;
    };
    fgScale: number;
    fgRotation: number;
  }